requirements.txt
ib_insync==0.9.86
aiosqlite==0.20.0
python-dotenv==1.0.1
aiogram==3.13.1
